applied config settings and also broke components into two various directories
