+++ resources/js/Components/Core/DangerButton.tsx
@@ -1,0 +1,22 @@
+import { ButtonHTMLAttributes } from 'react';
+
+export default function DangerButton({
+  className = '',
+  disabled,
+  children,
+  ...props
+}: ButtonHTMLAttributes<HTMLButtonElement>) {
+  return (
+    <button
+      {...props}
+      className={
+        `inline-flex items-center rounded-md border border-transparent bg-red-600 px-4 py-2 text-xs font-semibold uppercase tracking-widest text-white transition duration-150 ease-in-out hover:bg-red-500 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2 active:bg-red-700 dark:focus:ring-offset-gray-800 ${
+          disabled && 'opacity-25'
+        } ` + className
+      }
+      disabled={disabled}
+    >
+      {children}
+    </button>
+  );
+}
