+++ resources/js/Components/Core/InputLabel.tsx
@@ -1,0 +1,20 @@
+import { LabelHTMLAttributes } from 'react';
+
+export default function InputLabel({
+  value,
+  className = '',
+  children,
+  ...props
+}: LabelHTMLAttributes<HTMLLabelElement> & { value?: string }) {
+  return (
+    <label
+      {...props}
+      className={
+        `block text-sm font-medium text-gray-700 dark:text-gray-300 ` +
+        className
+      }
+    >
+      {value ? value : children}
+    </label>
+  );
+}
