+++ resources/js/Layouts/AuthenticatedLayout.tsx
@@ -1,7 +1,7 @@
-import ApplicationLogo from '@/Components/ApplicationLogo';
-import Dropdown from '@/Components/Dropdown';
-import NavLink from '@/Components/NavLink';
-import ResponsiveNavLink from '@/Components/ResponsiveNavLink';
+import ApplicationLogo from '@/Components/App/ApplicationLogo';
+import Dropdown from '@/Components/Core/Dropdown';
+import NavLink from '@/Components/Core/NavLink';
+import ResponsiveNavLink from '@/Components/Core/ResponsiveNavLink';
 import { Link, usePage } from '@inertiajs/react';
 import { PropsWithChildren, ReactNode, useState } from 'react';
