+++ resources/js/app.tsx
@@ -8,21 +8,21 @@ import { createRoot, hydrateRoot } from 'react-dom/client';
 const appName = import.meta.env.VITE_APP_NAME || 'Laravel';
 
 createInertiaApp({
-    title: (title) => `${title} - ${appName}`,
-    resolve: (name) =>
-        resolvePageComponent(
-            `./Pages/${name}.tsx`,
-            import.meta.glob('./Pages/**/*.tsx'),
-        ),
-    setup({ el, App, props }) {
-        if (import.meta.env.SSR) {
-            hydrateRoot(el, <App {...props} />);
-            return;
-        }
+  title: (title) => `${title} - ${appName}`,
+  resolve: (name) =>
+    resolvePageComponent(
+      `./Pages/${name}.tsx`,
+      import.meta.glob('./Pages/**/*.tsx'),
+    ),
+  setup({ el, App, props }) {
+    if (import.meta.env.SSR) {
+      hydrateRoot(el, <App {...props} />);
+      return;
+    }
 
-        createRoot(el).render(<App {...props} />);
-    },
-    progress: {
-        color: '#4B5563',
-    },
+    createRoot(el).render(<App {...props} />);
+  },
+  progress: {
+    color: '#4B5563',
+  },
 });
