+++ resources/js/ssr.tsx
@@ -8,26 +8,26 @@ import { route } from '../../vendor/tightenco/ziggy';
 const appName = import.meta.env.VITE_APP_NAME || 'Laravel';
 
 createServer((page) =>
-    createInertiaApp({
-        page,
-        render: ReactDOMServer.renderToString,
-        title: (title) => `${title} - ${appName}`,
-        resolve: (name) =>
-            resolvePageComponent(
-                `./Pages/${name}.tsx`,
-                import.meta.glob('./Pages/**/*.tsx'),
-            ),
-        setup: ({ App, props }) => {
-            /* eslint-disable */
-            // @ts-expect-error
-            global.route<RouteName> = (name, params, absolute) =>
-                route(name, params as any, absolute, {
-                    ...page.props.ziggy,
-                    location: new URL(page.props.ziggy.location),
-                });
-            /* eslint-enable */
+  createInertiaApp({
+    page,
+    render: ReactDOMServer.renderToString,
+    title: (title) => `${title} - ${appName}`,
+    resolve: (name) =>
+      resolvePageComponent(
+        `./Pages/${name}.tsx`,
+        import.meta.glob('./Pages/**/*.tsx'),
+      ),
+    setup: ({ App, props }) => {
+      /* eslint-disable */
+      // @ts-expect-error
+      global.route<RouteName> = (name, params, absolute) =>
+        route(name, params as any, absolute, {
+          ...page.props.ziggy,
+          location: new URL(page.props.ziggy.location),
+        });
+      /* eslint-enable */
 
-            return <App {...props} />;
-        },
-    }),
+      return <App {...props} />;
+    },
+  }),
 );
