+++ resources/js/types/global.d.ts
@@ -4,14 +4,14 @@ import { route as ziggyRoute } from 'ziggy-js';
 import { PageProps as AppPageProps } from './';
 
 declare global {
-    interface Window {
-        axios: AxiosInstance;
-    }
+  interface Window {
+    axios: AxiosInstance;
+  }
 
-    /* eslint-disable no-var */
-    var route: typeof ziggyRoute;
+  /* eslint-disable no-var */
+  var route: typeof ziggyRoute;
 }
 
 declare module '@inertiajs/core' {
-    interface PageProps extends InertiaPageProps, AppPageProps {}
+  interface PageProps extends InertiaPageProps, AppPageProps {}
 }
